requirements.txt
torch
torchvision
torchaudio
Pillow
opencv-python